requirements.txt
Flask==3.0.0
flask-socketio==5.3.5
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
PyPDF2==3.0.1
pdfplumber==0.10.3
Pillow>=12.0.0
python-multipart==0.0.6
Werkzeug==3.0.1
eventlet==0.33.3
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
google-cloud-firestore==2.16.1
google-auth==2.33.0
gunicorn==21.2.0

